Move W3C process events to separate package and classes
* Instead of having all the processor methods in the W3CActions class
create a new package called 'Processor' and have a separate class for
actions, key, pause, pointer processors and have a helper class
